Feat: calculate total credit

+++ Employee.h
@@ -57,7 +57,7 @@ class Employee {
     void checkIngredientInventory() const;
 
     // accessibly by supervisor role only
-    void restockIngredientInventory(int index, int quantity) const;
+    void restockIngredientInventory(int index, double quantity) const;
     void changeIngredientCost(int index, double cost);
     void addNewInventoryIngredientWeight(string name, double cost, double weight) const;
     void addNewInventoryIngredientPiece(string name, double cost, int piece) const;
